requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# LangChain and LangGraph - compatible versions
langchain==0.2.14
langgraph==0.1.19
langchain-community==0.2.12
langchain-ollama==0.1.3

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Data Analysis
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.28
requests==2.31.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Async and Caching
aiohttp==3.9.1
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced financial calculations
scipy==1.11.4
scikit-learn==1.3.2

# Local embeddings
sentence-transformers==2.2.2

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.4 